Update main configuration, disable some automations
This updates the main configuration to remove a dead ESP8266 project, and only send scripts to Alexa, for testing purposes. This also removes the now dead API Password auth, replaced by a long lived token for Alexa. Disabled the script automations that fire when the Chromecast is playing back, as I don't use the Cast function much anymore, and need to rebuild it with the Apple TV in mind.